Fix VR mode launch: disable guardian, remove crashing mesh manager, add data flush
- Remove PXR_SpatialMeshManager from scene (NullRef crash in InitMeshColor)
- Disable guardian system before enabling passthrough to prevent seethrough
  settings dialog from stealing focus on launch
- Add periodic CSV flush every 60 frames to prevent data loss
- Add startup diagnostics logging for XR state, devices, HMD, IMU
- Enable VR manifest declarations (pvr.app.type=vr, VRAPP category)
- Switch to WorldSpace canvas for VR rendering
- Add POV video recording via PXR_Enterprise.Record() with file discovery
- Enable XR auto-loading/running and passthrough in project settings

+++ SyncManager.cs
@@ -58,7 +58,7 @@ public class SyncManager : MonoBehaviour
             var ht = side == "left" ? HandType.HandLeft : HandType.HandRight;
             var jl = new HandJointLocations();
             if (!PXR_HandTracking.GetJointLocations(ht, ref jl) || jl.jointLocations == null || jl.jointLocations.Length < 1) return false;
-            if ((jl.jointLocations[0].locationStatus & (ulong)HandLocationStatus.PositionValid) == 0) return false;
+            if (((ulong)jl.jointLocations[0].locationStatus & (ulong)HandLocationStatus.PositionValid) == 0) return false;
             pos = jl.jointLocations[0].pose.Position.ToVector3();
             return true;
         }
